modified:   little_math_with_living_costs.py modified:   little_math_with_two_values.py

+++ little_math_with_living_costs.py
@@ -1,10 +1,10 @@
 # This program asks user about monthly income, how much school lunch costs, how many times user eats in campus and how much for the rest of living costs
 # after that, it prints yearly income, school lunch costs, monthly living costs and how much user has money left and how much money is saved in four years
 # if costs is cut down 20%
-monthly_income = int(input("How much is your monthly income? "))
-school_lunch_costs = int(input("How much school lunch costs? "))
-school_lunch_times = int(input("How many times do you eat at camput in one month? "))
-living_costs = int(input("How much rest of living costs is in one month? "))
+monthly_income = float(input("How much is your monthly income? "))
+school_lunch_costs = float(input("How much school lunch costs? "))
+school_lunch_times = float(input("How many times do you eat at camput in one month? "))
+living_costs = float(input("How much rest of living costs is in one month? "))
 print("Your yearly income is", monthly_income*12, "€")
 print("Monthly school lunch costs is", school_lunch_costs * school_lunch_times, "€")
 print("Monthly living costs (including school lunches) is", school_lunch_costs * school_lunch_times + living_costs, "€")
